refactor(font/glyph/sym): extract `mod tofu`
