+++ crates/monoxide-font/src/font/glyph/sym.rs
@@ -1,29 +1,10 @@
-use monoxide_script::{
-    ast::Glyph,
-    corner,
-    dsl::{IntoOutlineExt, SpiroBuilder},
-    let_settings,
-};
+mod tofu;
 
-use crate::{InputContext, font::dir::Alignment};
+use monoxide_script::prelude::*;
+
+pub use self::tofu::tofu;
+use crate::InputContext;
 
 pub fn space(_cx: &InputContext) -> Glyph {
     Glyph::builder().build()
-}
-
-pub fn tofu(cx: &InputContext) -> Glyph {
-    let_settings! { { cap, sbl, sbr, stw } = cx.settings(); }
-
-    Glyph::builder()
-        .outline(
-            SpiroBuilder::closed()
-                .insts([
-                    corner!(sbl, 0.).aligned(Alignment::Right),
-                    corner!(sbr, 0.),
-                    corner!(sbr, cap),
-                    corner!(sbl, cap),
-                ])
-                .stroked(stw),
-        )
-        .build()
 }
